+++ src/smt/proof_post_processor.cpp
@@ -887,7 +887,7 @@ Node ProofPostprocessCallback::expandMacros(PfRule id,
     builtin::BuiltinProofRuleChecker::getTheoryId(args[1], tid);
     MethodId mid = MethodId::RW_REWRITE;
     getMethodId(args[2], mid);
-    uint32_t recLimit = options().proof.proofRewriteRconsRecLimit;
+    int64_t recLimit = options().proof.proofRewriteRconsRecLimit;
     // attempt to reconstruct the proof of the equality into cdp using the
     // rewrite database proof reconstructor
     if (d_rdbPc.prove(cdp, args[0][0], args[0][1], tid, mid, recLimit))
